area as a method of Rectangle.
Methods are expected to take a reference to the object as first
parameter. They can borrow (self instead of &self) if methods are going to
alter the instance. This helps protecting the instance after the method
as altered it.

We do not need to specify the type for self, as that is implied here.
Rust knows this because it is an implantation of Rectangle